BROKEN Try moving the EditorProvider outside Layout
